# api/openapi.yaml
openapi: 3.0.0
info:
  title: Procedural Dungeon API
  version: 2.0.0
  description: REST API for the Procedural Dungeon roguelike game

servers:
  - url: http://localhost:4010
    description: Local development server

paths:
  /api/dungeon/generate:
    post:
      summary: Generate a new dungeon and start a game
      operationId: generateDungeon
      tags: [Dungeon]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                width:
                  type: integer
                  default: 40
                  minimum: 20
                  maximum: 80
                  description: Dungeon width in tiles
                height:
                  type: integer
                  default: 20
                  minimum: 10
                  maximum: 40
                  description: Dungeon height in tiles
              example:
                width: 40
                height: 20
      responses:
        '200':
          description: Game session created with generated dungeon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'

  /api/dungeon/{id}/state:
    get:
      summary: Get current game state
      operationId: getGameState
      tags: [Dungeon]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Session ID
      responses:
        '200':
          description: Current game state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStateResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/dungeon/{id}/move:
    post:
      summary: Move the player in the dungeon
      operationId: movePlayer
      tags: [Actions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [direction]
              properties:
                direction:
                  type: string
                  enum: [up, down, left, right]
              example:
                direction: up
      responses:
        '200':
          description: Move result with game state and events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/dungeon/{id}/pickup:
    post:
      summary: Pick up item at current player position
      operationId: pickupItem
      tags: [Actions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Pickup result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/dungeon/{id}/use-potion:
    post:
      summary: Use a health potion from inventory
      operationId: usePotion
      tags: [Actions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Potion used
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/dungeon/{id}/wait:
    post:
      summary: Wait one turn (monsters act)
      operationId: waitTurn
      tags: [Actions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Turn passed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStateResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/sessions:
    get:
      summary: List all active game sessions
      operationId: listSessions
      tags: [Sessions]
      responses:
        '200':
          description: Active sessions list
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionInfo'
                  count:
                    type: integer

  /api/sessions/{id}:
    delete:
      summary: Delete a game session
      operationId: deleteSession
      tags: [Sessions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session deleted
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    GameStateResponse:
      type: object
      properties:
        sessionId:
          type: string
        status:
          type: string
          enum: [PLAYING, VICTORY, DEFEAT]
        turn:
          type: integer
        remainingMonsters:
          type: integer
        player:
          $ref: '#/components/schemas/Player'
        dungeon:
          $ref: '#/components/schemas/Dungeon'
        events:
          type: array
          items:
            $ref: '#/components/schemas/GameEvent'
        messages:
          type: array
          items:
            type: string

    Player:
      type: object
      properties:
        x: { type: integer }
        y: { type: integer }
        hp: { type: integer }
        maxHp: { type: integer }
        atk: { type: integer }
        def: { type: integer }
        level: { type: integer }
        exp: { type: integer }
        expToNext: { type: integer }
        potions: { type: integer }
        weaponName: { type: string, nullable: true }

    Dungeon:
      type: object
      properties:
        width: { type: integer }
        height: { type: integer }
        tiles:
          type: array
          items:
            type: array
            items:
              $ref: '#/components/schemas/Tile'
        monsters:
          type: array
          items:
            $ref: '#/components/schemas/Monster'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'

    Tile:
      type: object
      properties:
        type: { type: string, enum: [FLOOR, WALL, WATER, CORRIDOR] }
        visible: { type: boolean }
        explored: { type: boolean }

    Monster:
      type: object
      properties:
        id: { type: integer }
        x: { type: integer }
        y: { type: integer }
        hp: { type: integer }
        maxHp: { type: integer }
        symbol: { type: string }

    Item:
      type: object
      properties:
        id: { type: integer }
        x: { type: integer }
        y: { type: integer }
        type: { type: string, enum: [WEAPON, POTION] }
        name: { type: string }
        symbol: { type: string }

    GameEvent:
      type: object
      properties:
        type: { type: string }
        message: { type: string }

    SessionInfo:
      type: object
      properties:
        sessionId: { type: string }
        createdAt: { type: string, format: date-time }
        lastActiveAt: { type: string, format: date-time }
        turn: { type: integer }
        status: { type: string }

    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        message: { type: string }

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Session not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

tags:
  - name: Dungeon
    description: Dungeon generation and state
  - name: Actions
    description: Player actions
  - name: Sessions
    description: Session management